+++ doc/shared/Changelog.rst
@@ -29,6 +29,19 @@ Changelog
 Changes to SUNDIALS in release X.Y.Z
 ====================================
 
+.. SED_REPLACEMENT_KEY
+
+.. _Changelog.7.7.0:
+
+Changes to SUNDIALS in release 7.7.0
+====================================
+
+**Bug Fixes**
+
+Fixed a minor bug where the number of required stages for STS methods 
+in the LSRKStep module was incorrectly computed using the spectral 
+radius instead of the real part of the Jacobian eigenvalues.
+
 .. include:: RecentChanges_link.rst
 
 .. _Changelog.7.6.0: